Refactor classes and methods for better clarity
