+++ cantor-grpc-client/src/main/java/com/salesforce/cantor/grpc/CantorOnGrpc.java
@@ -5,7 +5,7 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-package com.salesforce.cantor.grpc.user;
+package com.salesforce.cantor.grpc;
 
 import com.salesforce.cantor.*;
 import com.salesforce.cantor.management.CantorCredentials;
