+++ cantor-grpc-service/src/main/java/com/salesforce/cantor/grpc/EventsGrpcService.java
@@ -5,7 +5,7 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-package com.salesforce.cantor.grpc.open;
+package com.salesforce.cantor.grpc;
 
 import com.google.protobuf.ByteString;
 import com.salesforce.cantor.Cantor;
@@ -18,7 +18,7 @@ import java.io.IOException;
 import java.util.*;
 
 import static com.salesforce.cantor.common.CommonPreconditions.checkArgument;
-import static com.salesforce.cantor.grpc.open.GrpcUtils.*;
+import static com.salesforce.cantor.grpc.GrpcUtils.*;
 
 public class EventsGrpcService extends EventsServiceGrpc.EventsServiceImplBase {
