+++ cantor-grpc-service/src/main/java/com/salesforce/cantor/grpc/GrpcUtils.java
@@ -5,7 +5,7 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-package com.salesforce.cantor.grpc.open;
+package com.salesforce.cantor.grpc;
 
 import com.google.protobuf.Message;
 import io.grpc.Status;
