+++ cantor-grpc-service/src/main/java/com/salesforce/cantor/grpc/auth/AuthorizedObjectsGrpcService.java
@@ -9,11 +9,11 @@ package com.salesforce.cantor.grpc.auth;
 
 import com.salesforce.cantor.Cantor;
 import com.salesforce.cantor.grpc.objects.*;
-import com.salesforce.cantor.grpc.open.ObjectsGrpcService;
+import com.salesforce.cantor.grpc.ObjectsGrpcService;
 import io.grpc.stub.StreamObserver;
 
 import static com.salesforce.cantor.grpc.auth.UserUtils.*;
-import static com.salesforce.cantor.grpc.open.GrpcUtils.*;
+import static com.salesforce.cantor.grpc.GrpcUtils.*;
 
 public class AuthorizedObjectsGrpcService extends ObjectsGrpcService {
 
@@ -28,7 +28,7 @@ public class AuthorizedObjectsGrpcService extends ObjectsGrpcService {
 
     @Override
     public void create(final CreateRequest request, final StreamObserver<VoidResponse> responseObserver) {
-        if (writeRequestInvalid(request.getNamespace())) {
+        if (!writeRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -37,7 +37,7 @@ public class AuthorizedObjectsGrpcService extends ObjectsGrpcService {
 
     @Override
     public void drop(final DropRequest request, final StreamObserver<VoidResponse> responseObserver) {
-        if (writeRequestInvalid(request.getNamespace())) {
+        if (!writeRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -46,7 +46,7 @@ public class AuthorizedObjectsGrpcService extends ObjectsGrpcService {
 
     @Override
     public void keys(final KeysRequest request, final StreamObserver<KeysResponse> responseObserver) {
-        if (readRequestInvalid(request.getNamespace())) {
+        if (!readRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -55,7 +55,7 @@ public class AuthorizedObjectsGrpcService extends ObjectsGrpcService {
 
     @Override
     public void get(final GetRequest request, final StreamObserver<GetResponse> responseObserver) {
-        if (readRequestInvalid(request.getNamespace())) {
+        if (!readRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -64,7 +64,7 @@ public class AuthorizedObjectsGrpcService extends ObjectsGrpcService {
 
     @Override
     public void store(final StoreRequest request, final StreamObserver<VoidResponse> responseObserver) {
-        if (writeRequestInvalid(request.getNamespace())) {
+        if (!writeRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -73,7 +73,7 @@ public class AuthorizedObjectsGrpcService extends ObjectsGrpcService {
 
     @Override
     public void delete(final DeleteRequest request, final StreamObserver<DeleteResponse> responseObserver) {
-        if (writeRequestInvalid(request.getNamespace())) {
+        if (!writeRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -82,7 +82,7 @@ public class AuthorizedObjectsGrpcService extends ObjectsGrpcService {
 
     @Override
     public void size(final SizeRequest request, final StreamObserver<SizeResponse> responseObserver) {
-        if (readRequestInvalid(request.getNamespace())) {
+        if (!readRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
