+++ cantor-grpc-service/src/main/java/com/salesforce/cantor/grpc/auth/AuthorizedSetsGrpcService.java
@@ -8,12 +8,12 @@
 package com.salesforce.cantor.grpc.auth;
 
 import com.salesforce.cantor.Cantor;
-import com.salesforce.cantor.grpc.open.SetsGrpcService;
+import com.salesforce.cantor.grpc.SetsGrpcService;
 import com.salesforce.cantor.grpc.sets.*;
 import io.grpc.stub.StreamObserver;
 
 import static com.salesforce.cantor.grpc.auth.UserUtils.*;
-import static com.salesforce.cantor.grpc.open.GrpcUtils.*;
+import static com.salesforce.cantor.grpc.GrpcUtils.*;
 
 public class AuthorizedSetsGrpcService extends SetsGrpcService {
 
@@ -28,7 +28,7 @@ public class AuthorizedSetsGrpcService extends SetsGrpcService {
 
     @Override
     public void create(final CreateRequest request, final StreamObserver<VoidResponse> responseObserver) {
-        if (writeRequestInvalid(request.getNamespace())) {
+        if (!writeRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -37,7 +37,7 @@ public class AuthorizedSetsGrpcService extends SetsGrpcService {
 
     @Override
     public void drop(final DropRequest request, final StreamObserver<VoidResponse> responseObserver) {
-        if (writeRequestInvalid(request.getNamespace())) {
+        if (!writeRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -46,7 +46,7 @@ public class AuthorizedSetsGrpcService extends SetsGrpcService {
 
     @Override
     public void get(final GetRequest request, final StreamObserver<GetResponse> responseObserver) {
-        if (readRequestInvalid(request.getNamespace())) {
+        if (!readRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -55,7 +55,7 @@ public class AuthorizedSetsGrpcService extends SetsGrpcService {
 
     @Override
     public void union(final UnionRequest request, final StreamObserver<UnionResponse> responseObserver) {
-        if (readRequestInvalid(request.getNamespace())) {
+        if (!readRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -64,7 +64,7 @@ public class AuthorizedSetsGrpcService extends SetsGrpcService {
 
     @Override
     public void intersect(final IntersectRequest request, final StreamObserver<IntersectResponse> responseObserver) {
-        if (readRequestInvalid(request.getNamespace())) {
+        if (!readRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -73,7 +73,7 @@ public class AuthorizedSetsGrpcService extends SetsGrpcService {
 
     @Override
     public void pop(final PopRequest request, final StreamObserver<PopResponse> responseObserver) {
-        if (writeRequestInvalid(request.getNamespace())) {
+        if (!writeRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -82,7 +82,7 @@ public class AuthorizedSetsGrpcService extends SetsGrpcService {
 
     @Override
     public void add(final AddRequest request, final StreamObserver<VoidResponse> responseObserver) {
-        if (writeRequestInvalid(request.getNamespace())) {
+        if (!writeRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -91,7 +91,7 @@ public class AuthorizedSetsGrpcService extends SetsGrpcService {
 
     @Override
     public void addBatch(final AddBatchRequest request, final StreamObserver<VoidResponse> responseObserver) {
-        if (writeRequestInvalid(request.getNamespace())) {
+        if (!writeRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -100,7 +100,7 @@ public class AuthorizedSetsGrpcService extends SetsGrpcService {
 
     @Override
     public void deleteBetween(final DeleteBetweenRequest request, final StreamObserver<VoidResponse> responseObserver) {
-        if (writeRequestInvalid(request.getNamespace())) {
+        if (!writeRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -109,7 +109,7 @@ public class AuthorizedSetsGrpcService extends SetsGrpcService {
 
     @Override
     public void deleteEntry(final DeleteEntryRequest request, final StreamObserver<DeleteEntryResponse> responseObserver) {
-        if (writeRequestInvalid(request.getNamespace())) {
+        if (!writeRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -118,7 +118,7 @@ public class AuthorizedSetsGrpcService extends SetsGrpcService {
 
     @Override
     public void deleteBatch(final DeleteBatchRequest request, final StreamObserver<VoidResponse> responseObserver) {
-        if (writeRequestInvalid(request.getNamespace())) {
+        if (!writeRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -127,7 +127,7 @@ public class AuthorizedSetsGrpcService extends SetsGrpcService {
 
     @Override
     public void keys(final KeysRequest request, final StreamObserver<KeysResponse> responseObserver) {
-        if (readRequestInvalid(request.getNamespace())) {
+        if (!readRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -136,7 +136,7 @@ public class AuthorizedSetsGrpcService extends SetsGrpcService {
 
     @Override
     public void sets(final SetsRequest request, final StreamObserver<SetsResponse> responseObserver) {
-        if (readRequestInvalid(request.getNamespace())) {
+        if (!readRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -145,7 +145,7 @@ public class AuthorizedSetsGrpcService extends SetsGrpcService {
 
     @Override
     public void size(final SizeRequest request, final StreamObserver<SizeResponse> responseObserver) {
-        if (readRequestInvalid(request.getNamespace())) {
+        if (!readRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -154,7 +154,7 @@ public class AuthorizedSetsGrpcService extends SetsGrpcService {
 
     @Override
     public void weight(final WeightRequest request, final StreamObserver<WeightResponse> responseObserver) {
-        if (readRequestInvalid(request.getNamespace())) {
+        if (!readRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
@@ -163,7 +163,7 @@ public class AuthorizedSetsGrpcService extends SetsGrpcService {
 
     @Override
     public void inc(final IncRequest request, final StreamObserver<IncResponse> responseObserver) {
-        if (writeRequestInvalid(request.getNamespace())) {
+        if (!writeRequestValid(request.getNamespace())) {
             sendError(responseObserver, new UnauthorizedException(request.toString()));
             return;
         }
